update client page finished

+++ src/clients/hooks/useCreateClientHook.ts
@@ -7,12 +7,25 @@ import {
 } from "@/src/utils";
 import { ChangeEvent, FormEvent, useEffect, useState } from "react";
 import { validateCreateClientData } from "../utils/validateCreateClientData";
-import { createClientAction } from "../actions/clientActions";
-import { setCookie } from "cookies-next";
-import { CLIENT_LABELS, COOKIE_CLIENT_IMAGES } from "../utils/const";
-import { ImagesToCreate } from "../interfaces/client-to-create.interface";
 
-const INITIAL_STATE = {
+import { setCookie, deleteCookie } from "cookies-next";
+import {
+  CLIENT_LABELS,
+  COOKIE_CLIENT_IMAGES,
+  COOKIE_UPDATE_CLIENT_DELETE_IMAGES,
+  COOKIE_UPDATE_CLIENT_IMAGES
+} from "../utils/const";
+import {
+  IClientToCreate,
+  ImagesToCreate
+} from "../interfaces/client-to-create.interface";
+import {
+  ICreateClientAction,
+  IUpdateClientAction
+} from "../interfaces/actions.interface";
+import { useRouter } from "next/navigation";
+
+const INITIAL_STATE: IClientToCreate = {
   name: "",
   lastName: "",
   phoneNumber1: "",
@@ -29,10 +42,26 @@ const INITIAL_STATE = {
 
 interface Props {
   initialImages: ImagesToCreate[];
+  client?: IClientToCreate;
+  action: ICreateClientAction | IUpdateClientAction;
+  type: "create" | "update";
+  id?: string;
+  initialImagesToDelete?: string[];
 }
-export const useCreateClientHook = ({ initialImages = [] }: Props) => {
-  const [formValues, setFormValues] = useState(INITIAL_STATE);
+
+export const useCreateClientHook = ({
+  initialImages = [],
+  client = INITIAL_STATE,
+  action,
+  type,
+  id,
+  initialImagesToDelete = []
+}: Props) => {
+  const [formValues, setFormValues] = useState(client);
   const [images, setImages] = useState<ImagesToCreate[]>(initialImages);
+  const [imagesToDeleteWhenUpdate, setImagesToDeleteWhenUpdate] = useState<
+    string[]
+  >(initialImagesToDelete);
   const showNotification = useNotificationStore(
     state => state.showNotification
   );
@@ -41,6 +70,7 @@ export const useCreateClientHook = ({ initialImages = [] }: Props) => {
 
   const [loadingImages, setLoadingImages] = useState(false);
 
+  const router = useRouter();
   const handleAddImages = async ({
     target: { files, name }
   }: ChangeEvent<HTMLInputElement>) => {
@@ -63,22 +93,27 @@ export const useCreateClientHook = ({ initialImages = [] }: Props) => {
   };
 
   const handleDeleteImages = async (url: string) => {
-    const key = url.split(".com/")[1];
+    if (type === "create") {
+      const key = url.split(".com/")[1];
 
-    try {
-      const response = await fetch(`/api/deleteImage?key=${key}`, {
-        method: "DELETE"
-      });
-      await response.json();
+      try {
+        const response = await fetch(`/api/deleteImage?key=${key}`, {
+          method: "DELETE"
+        });
+        await response.json();
 
-      const newImages = images.filter(image => image.url !== url);
+        const newImages = images.filter(image => image.url !== url);
 
-      setImages(newImages);
-    } catch (error) {
-      showNotification({
-        type: "error",
-        text: GENERAL_LABELS.IMAGES.ERROR.DELETE_IMAGE
-      });
+        setImages(newImages);
+      } catch (error) {
+        showNotification({
+          type: "error",
+          text: GENERAL_LABELS.IMAGES.ERROR.DELETE_IMAGE
+        });
+      }
+    } else {
+      setImagesToDeleteWhenUpdate([...imagesToDeleteWhenUpdate, url]);
+      setImages(images.filter(img => img.url !== url));
     }
   };
 
@@ -105,34 +140,91 @@ export const useCreateClientHook = ({ initialImages = [] }: Props) => {
 
     setLoading(true);
 
-    console.log(formValues);
-
-    const { data, error, success } = await createClientAction({
-      ...formValues,
-      images
-    });
-
-    if (error) {
-      showNotification({
-        text: GENERAL_LABELS.ERRORS.NOTIFICATION_ERROR,
-        type: "error"
+    if (type === "create") {
+      const { data, error, success, message } = await (
+        action as ICreateClientAction
+      )({
+        ...formValues,
+        images
       });
-    } else if (success && data) {
-      showNotification({
-        text: CLIENT_LABELS.NOTIFICATIONS.CLIENT_CREATED,
-        type: "success"
+
+      if (error) {
+        showNotification({
+          text: `${GENERAL_LABELS.ERRORS.NOTIFICATION_ERROR}: ${message}`,
+          type: "error"
+        });
+      } else if (success && data) {
+        showNotification({
+          text: CLIENT_LABELS.NOTIFICATIONS.CLIENT_CREATED,
+          type: "success"
+        });
+        setFormValues(INITIAL_STATE);
+      }
+
+      setImages([]);
+    } else if (type === "update") {
+      const { data, error, success, message } = await (
+        action as IUpdateClientAction
+      )(id!, {
+        ...formValues,
+        images
       });
-      setFormValues(INITIAL_STATE);
+
+      for (const imageToDelete of imagesToDeleteWhenUpdate) {
+        const key = imageToDelete.split(".com/")[1];
+
+        try {
+          await fetch(`/api/deleteImage?key=${key}`, {
+            method: "DELETE"
+          });
+        } catch (error) {
+          showNotification({
+            type: "error",
+            text: GENERAL_LABELS.IMAGES.ERROR.DELETE_IMAGE
+          });
+        }
+      }
+
+      if (error) {
+        showNotification({
+          text: `${GENERAL_LABELS.ERRORS.NOTIFICATION_ERROR}: ${message}`,
+          type: "error"
+        });
+      } else if (success && data) {
+        deleteCookie(COOKIE_UPDATE_CLIENT_IMAGES);
+        deleteCookie(COOKIE_UPDATE_CLIENT_DELETE_IMAGES);
+        showNotification({
+          text: CLIENT_LABELS.NOTIFICATIONS.CLIENT_UPDATED,
+          type: "success"
+        });
+      }
     }
 
-    setImages([]);
     setLoading(false);
     setShowErrors(false);
+    router.refresh();
   };
 
   useEffect(() => {
-    setCookie(COOKIE_CLIENT_IMAGES, JSON.stringify(images));
-  }, [images]);
+    if (type === "update") {
+      setCookie(`${COOKIE_UPDATE_CLIENT_IMAGES}-${id}`, JSON.stringify(images));
+    }
+  }, [images, id, type]);
+
+  useEffect(() => {
+    if (type === "update") {
+      setCookie(
+        `${COOKIE_UPDATE_CLIENT_DELETE_IMAGES}-${id}`,
+        JSON.stringify(imagesToDeleteWhenUpdate)
+      );
+    }
+  }, [imagesToDeleteWhenUpdate, id, type]);
+
+  useEffect(() => {
+    if (type === "create") {
+      setCookie(COOKIE_CLIENT_IMAGES, JSON.stringify(images));
+    }
+  }, [images, type]);
 
   return {
     showErrors,
